pine64-rock64: Apply power to USB ports
They end up working better when power is provided.

How does this make it work?

 1. Make the DWC3 driver work with rk3328 (by stating it will)
 2. Add the necessary supply nodes to the DT

Note that the innate FDT Linux will not enable USB ports. So while Linux
will boot, it will likely not be able to mount what's on the USB stick,
unless it is loading another FDT.

This is because the U-Boot FDT conflicts with the upstream drivers.

To work around the issue, use any upstream method to load another FDT,
some listed here:

 - https://source.denx.de/u-boot/u-boot/-/blob/v2022.04/doc/develop/distro.rst

Since the issue stems from the innate FDT being the U-Boot-modified FDT,
the core issue is not specific to the Rock64 board. A generic solution
will need to be explored, as using the innate FDT may not work as
expected.