radxa-rock5b: Preliminary build from vendor BSP
Though this is preliminary, this build includes an interesting
opinionated change: proprietary DDR init that is patched to set the
serial to 115200.

This approach will be used on other boards where there is a need for the
proprietary RAM init too.

**NOTE**: This does not actually build as-is; rockchip-specific patches
were disabled hapazardly on my end to make it build, they will be fixed
at some point.